test: eliminate console errors and warnings in unit tests
